use stimulus to show google maps
