+++ app/javascript/application.js
@@ -4,3 +4,12 @@
 //= require bootstrap-sprockets
 import "@hotwired/turbo-rails";
 import "controllers";
+
+window.initMap = () => {
+  console.log("initMap was called");
+  const event = new CustomEvent("map-loaded", {
+    bubbles: true,
+    cancelable: true,
+  });
+  window.dispatchEvent(event);
+};
