+++ app/views/home/search.html.erb
@@ -1,11 +1,9 @@
 <% content_for :meta do %>
   <meta name="turbo-cache-control" content="no-cache">
 <% end %>
-<% content_for :js do %>
-  <%= javascript_import_module_tag "map" %>
-<% end %>
 <div id="loading-spinner" class="loading-spinner">
   <div class="spinner"></div>
 </div>
-<div id="map"></div>
-<script src="https://maps.googleapis.com/maps/api/js?key=<%= ENV['GOOGLE_MAP_API_KEY'] %>&callback=initMap" defer></script>
+<div data-controller="map" data-action="map-loaded@window->map#initializeMap">
+  <div id="map" data-map-target="map"></div>
+</div>
